"LA.QSSI.12.0.r1-08200-qssi.0"

+++ services/core/java/com/android/server/am/ActiveServices.java
@@ -306,6 +306,9 @@ public final class ActiveServices {
 
     private IServicetracker mServicetracker;
 
+    private final boolean isLowRamDevice =
+            SystemProperties.getBoolean("ro.config.low_ram", false);
+
     String mLastAnrDump;
 
     AppWidgetManagerInternal mAppWidgetManagerInternal;
@@ -2848,28 +2851,30 @@ public final class ActiveServices {
             }
             clist.add(c);
 
-            ServiceData sData = new ServiceData();
-            sData.packageName = s.packageName;
-            sData.processName = s.shortInstanceName;
-            sData.lastActivity = s.lastActivity;
-            if (s.app != null) {
-                sData.pid = s.app.getPid();
-                sData.serviceB = s.app.mState.isServiceB();
-            } else {
-                 sData.pid = -1;
-                 sData.serviceB = false;
-            }
-
-            ClientData cData = new ClientData();
-            cData.processName = callerApp.processName;
-            cData.pid = callerApp.getPid();
-            try {
-                if (getServicetrackerInstance()) {
-                   mServicetracker.bindService(sData, cData);
+            if (!isLowRamDevice) {
+                ServiceData sData = new ServiceData();
+                sData.packageName = s.packageName;
+                sData.processName = s.shortInstanceName;
+                sData.lastActivity = s.lastActivity;
+                if (s.app != null) {
+                    sData.pid = s.app.getPid();
+                    sData.serviceB = s.app.mState.isServiceB();
+                } else {
+                    sData.pid = -1;
+                    sData.serviceB = false;
                 }
-            } catch (RemoteException e) {
-                Slog.e(TAG, "Failed to send bind details to servicetracker HAL", e);
-                mServicetracker = null;
+
+                ClientData cData = new ClientData();
+                cData.processName = callerApp.processName;
+                cData.pid = callerApp.getPid();
+                try {
+                    if (getServicetrackerInstance()) {
+                        mServicetracker.bindService(sData, cData);
+                    }
+                } catch (RemoteException e) {
+                    Slog.e(TAG, "Failed to send bind details to servicetracker HAL", e);
+                    mServicetracker = null;
+                }
             }
 
             boolean needOomAdj = false;
@@ -3054,28 +3059,30 @@ public final class ActiveServices {
         try {
             while (clist.size() > 0) {
                 ConnectionRecord r = clist.get(0);
-                ServiceData sData = new ServiceData();
-                sData.packageName = r.binding.service.packageName;
-                sData.processName = r.binding.service.shortInstanceName;
-                sData.lastActivity = r.binding.service.lastActivity;
-                if(r.binding.service.app != null) {
-                    sData.pid = r.binding.service.app.getPid();
-                    sData.serviceB = r.binding.service.app.mState.isServiceB();
-                } else {
-                    sData.pid = -1;
-                    sData.serviceB = false;
-                }
-
-                ClientData cData = new ClientData();
-                cData.processName = r.binding.client.processName;
-                cData.pid = r.binding.client.getPid();
-                try {
-                    if (getServicetrackerInstance()) {
-                        mServicetracker.unbindService(sData, cData);
+                if (!isLowRamDevice) {
+                    ServiceData sData = new ServiceData();
+                    sData.packageName = r.binding.service.packageName;
+                    sData.processName = r.binding.service.shortInstanceName;
+                    sData.lastActivity = r.binding.service.lastActivity;
+                    if(r.binding.service.app != null) {
+                        sData.pid = r.binding.service.app.getPid();
+                        sData.serviceB = r.binding.service.app.mState.isServiceB();
+                    } else {
+                        sData.pid = -1;
+                        sData.serviceB = false;
                     }
-                } catch (RemoteException e) {
-                    Slog.e(TAG, "Failed to send unbind details to servicetracker HAL", e);
-                    mServicetracker = null;
+
+                    ClientData cData = new ClientData();
+                    cData.processName = r.binding.client.processName;
+                    cData.pid = r.binding.client.getPid();
+                    try {
+                        if (getServicetrackerInstance()) {
+                            mServicetracker.unbindService(sData, cData);
+                        }
+                    } catch (RemoteException e) {
+                        Slog.e(TAG, "Failed to send unbind details to servicetracker HAL", e);
+                        mServicetracker = null;
+                    }
                 }
                 removeConnectionLocked(r, null, null, true);
                 if (clist.size() > 0 && clist.get(0) == r) {
@@ -4086,20 +4093,22 @@ public final class ActiveServices {
             r.postNotification();
             created = true;
 
-            ServiceData sData = new ServiceData();
-            sData.packageName = r.packageName;
-            sData.processName = r.shortInstanceName;
-            sData.pid = r.app.getPid();
-            sData.lastActivity = r.lastActivity;
-            sData.serviceB = r.app.mState.isServiceB();
+            if (!isLowRamDevice) {
+                ServiceData sData = new ServiceData();
+                sData.packageName = r.packageName;
+                sData.processName = r.shortInstanceName;
+                sData.pid = r.app.getPid();
+                sData.lastActivity = r.lastActivity;
+                sData.serviceB = r.app.mState.isServiceB();
 
-            try {
-                if (getServicetrackerInstance()) {
-                    mServicetracker.startService(sData);
+                try {
+                    if (getServicetrackerInstance()) {
+                        mServicetracker.startService(sData);
+                    }
+                } catch (RemoteException e) {
+                    Slog.e(TAG, "Failed to send start details to servicetracker HAL", e);
+                    mServicetracker = null;
                 }
-            } catch (RemoteException e) {
-                Slog.e(TAG, "Failed to send start details to servicetracker HAL", e);
-                mServicetracker = null;
             }
         } catch (DeadObjectException e) {
             Slog.w(TAG, "Application dead when creating service " + r);
@@ -4300,24 +4309,26 @@ public final class ActiveServices {
     private void bringDownServiceLocked(ServiceRecord r, boolean enqueueOomAdj) {
         //Slog.i(TAG, "Bring down service:");
         //r.dump("  ");
-        ServiceData sData = new ServiceData();
-        sData.packageName = r.packageName;
-        sData.processName = r.shortInstanceName;
-        sData.lastActivity = r.lastActivity;
-        if (r.app != null) {
-            sData.pid = r.app.getPid();
-        } else {
-            sData.pid = -1;
-            sData.serviceB = false;
-        }
-
-        try {
-            if (getServicetrackerInstance()) {
-                mServicetracker.destroyService(sData);
+        if (!isLowRamDevice) {
+            ServiceData sData = new ServiceData();
+            sData.packageName = r.packageName;
+            sData.processName = r.shortInstanceName;
+            sData.lastActivity = r.lastActivity;
+            if (r.app != null) {
+                sData.pid = r.app.getPid();
+            } else {
+                sData.pid = -1;
+                sData.serviceB = false;
             }
-        } catch (RemoteException e) {
-            Slog.e(TAG, "Failed to send destroy details to servicetracker HAL", e);
-            mServicetracker = null;
+
+            try {
+                if (getServicetrackerInstance()) {
+                    mServicetracker.destroyService(sData);
+                }
+            } catch (RemoteException e) {
+                Slog.e(TAG, "Failed to send destroy details to servicetracker HAL", e);
+                mServicetracker = null;
+            }
         }
         // Report to all of the connections that the service is no longer
         // available.
@@ -5065,7 +5076,7 @@ public final class ActiveServices {
         }
 
         try {
-            if (getServicetrackerInstance()) {
+            if (!isLowRamDevice && getServicetrackerInstance()) {
                 mServicetracker.killProcess(app.getPid());
             }
         } catch (RemoteException e) {
